Overview of JC CHEMICAL SYSTEM
JC CHEMICAL SYSTEM is a desktop application system developed for Mr. Justice Cudjoe a licensed chemist in Malejor, Accra in Ghana. It has an inventory and a point of sale. The inventory will help him keep track of his stocks and the point of sale will help him perform his daily transactions in an effective and efficient way. He will be able to print receipts after the sales and transactions made.